requirements.txt
absl-py==0.8.0
astor==0.8.0
boto3==1.12.28
botocore==1.15.28
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.1
docutils==0.15.2
filelock==3.0.12
Flask==1.1.1
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.4
oauthlib==3.1.0
opt-einsum==3.0.1
pandas==1.0.3
protobuf==3.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scipy==1.4.1
sentencepiece==0.1.85
six==1.12.0
termcolor==1.1.0
tokenizers==0.5.2
torch==1.4.0
tqdm==4.36.1
transformers==2.6.0
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
